pyproject.toml
[project]
name = "learnease-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "TranVanTienDat",email = "ckim32993@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "flask-restful (>=0.3.10,<0.4.0)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.22,<0.4.0)",
    "openai (>=1.76.0,<2.0.0)",
    "langchain-openai (>=0.3.14,<0.4.0)",
    "langchain-together (>=0.3.0,<0.4.0)",
    "langchain-google-genai (>=2.1.4,<3.0.0)",
    "langgraph (>=0.4.1,<0.5.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
